Change private vars to public, added timer on cycle
When opcodes are based back the easiest way for the cpu to modify its
own memory, stack and process counter was to attach them to the object
so i could access them with 'this'. Maybe not the greatest but we will
see what happens.

On emulate cycle we minus timers if they arent zero